refactor(core): Factor out object validation into a shared helper + leverage in ValidateBody hook
