+++ packages/core/src/common/validation/helpers/index.ts
@@ -1,1 +1,3 @@
-export { isFunction } from './is-function.util';
+export { createObjectValidator } from "./validate-object.util";
+export { extractProperties } from "./extract-properties.util";
+export { isFunction } from "./is-function.util";
